Multithreaded execution is much simpler
ThreadPool now simply executes blocks off a queue and doesn't return
futures. It does collect exceptions which are thrown in ThreadPool#join

Task#invoke_wtih_call_chain takes a new parameter which indicates if
the caller wants to bail if it can't get the lock.

Task#invoke_prerequisites will now queue the prerequisites to be
executed on the the ThreadPool queue, but by passing false to
Task#invoke_with_call_chain, the threads in the pool will not wait on
the task if they can't get the lock. In either the MultiTask or Task
case, Task#invoke_prerequisites will still continue executing the
prerequisites (and waiting) in the current thread, but it's very likely
they will already have been processed by the time the current thread
gets to them.

This has the effect of getting the threads in the pool to crawl the
task tree like little ants, executing what they can and bailing on
what's already claimed by another thread. The current thread marches
on, making sure the entire tree is finished before returning.

Removed Promise.rb as it is no longer needed.